requirements.txt
pinecone
openai
python-dotenv
pyTelegramBotAPI
haystack-ai
pinecone-haystack
requests
docling
docling-core[chunking]